+++ libraries/common/src/main/java/androidx/media3/common/MediaItem.java
@@ -1179,7 +1179,7 @@ public final class MediaItem implements Bundleable {
         @Nullable Object tag,
         long imageDurationMs) {
       this.uri = uri;
-      this.mimeType = mimeType;
+      this.mimeType = MimeTypes.normalizeMimeType(mimeType);
       this.drmConfiguration = drmConfiguration;
       this.adsConfiguration = adsConfiguration;
       this.streamKeys = streamKeys;
@@ -1612,7 +1612,7 @@ public final class MediaItem implements Bundleable {
       /** Sets the MIME type. */
       @CanIgnoreReturnValue
       public Builder setMimeType(@Nullable String mimeType) {
-        this.mimeType = mimeType;
+        this.mimeType = MimeTypes.normalizeMimeType(mimeType);
         return this;
       }
 
@@ -1694,7 +1694,7 @@ public final class MediaItem implements Bundleable {
         @Nullable String label,
         @Nullable String id) {
       this.uri = uri;
-      this.mimeType = mimeType;
+      this.mimeType = MimeTypes.normalizeMimeType(mimeType);
       this.language = language;
       this.selectionFlags = selectionFlags;
       this.roleFlags = roleFlags;