+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/offline/DownloadRequest.java
@@ -23,6 +23,7 @@ import android.os.Parcelable;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.MediaItem;
+import androidx.media3.common.MimeTypes;
 import androidx.media3.common.StreamKey;
 import androidx.media3.common.util.Assertions;
 import androidx.media3.common.util.UnstableApi;
@@ -61,7 +62,7 @@ public final class DownloadRequest implements Parcelable {
     /** Sets the {@link DownloadRequest#mimeType}. */
     @CanIgnoreReturnValue
     public Builder setMimeType(@Nullable String mimeType) {
-      this.mimeType = mimeType;
+      this.mimeType = MimeTypes.normalizeMimeType(mimeType);
       return this;
     }