+++ libraries/extractor/src/main/java/androidx/media3/extractor/metadata/id3/Id3Decoder.java
@@ -18,6 +18,7 @@ package androidx.media3.extractor.metadata.id3;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.Metadata;
+import androidx.media3.common.MimeTypes;
 import androidx.media3.common.util.Log;
 import androidx.media3.common.util.ParsableBitArray;
 import androidx.media3.common.util.ParsableByteArray;
@@ -557,7 +558,8 @@ public final class Id3Decoder extends SimpleMetadataDecoder {
     id3Data.readBytes(data, 0, frameSize - 1);
 
     int mimeTypeEndIndex = indexOfZeroByte(data, 0);
-    String mimeType = new String(data, 0, mimeTypeEndIndex, Charsets.ISO_8859_1);
+    String mimeType =
+        MimeTypes.normalizeMimeType(new String(data, 0, mimeTypeEndIndex, Charsets.ISO_8859_1));
 
     int filenameStartIndex = mimeTypeEndIndex + 1;
     int filenameEndIndex = indexOfTerminator(data, filenameStartIndex, encoding);