+++ libraries/transformer/src/main/java/androidx/media3/transformer/Transformer.java
@@ -174,6 +174,7 @@ public final class Transformer {
      */
     @CanIgnoreReturnValue
     public Builder setAudioMimeType(String audioMimeType) {
+      audioMimeType = MimeTypes.normalizeMimeType(audioMimeType);
       checkArgument(MimeTypes.isAudio(audioMimeType), "Not an audio MIME type: " + audioMimeType);
       this.audioMimeType = audioMimeType;
       return this;
@@ -205,6 +206,7 @@ public final class Transformer {
      */
     @CanIgnoreReturnValue
     public Builder setVideoMimeType(String videoMimeType) {
+      videoMimeType = MimeTypes.normalizeMimeType(videoMimeType);
       checkArgument(MimeTypes.isVideo(videoMimeType), "Not a video MIME type: " + videoMimeType);
       this.videoMimeType = videoMimeType;
       return this;