+++ RELEASENOTES.md
@@ -7,6 +7,8 @@
         resource URIs where `package` is different to the package of the current
         application. This has always been documented to work, but wasn't
         correctly implemented until now.
+    *   Normalize MIME types set by app code or read from media to be fully
+        lower-case.
 *   ExoPlayer:
     *   Add `PreloadMediaSource` and `PreloadMediaPeriod` that allows apps to
         preload the media source at a specific start position before playback,